+++ b3lb/rest/classes/api.py
@@ -33,7 +33,7 @@ from re import compile, escape
 from requests import get
 from rest.b3lb.metrics import incr_metric, update_create_metrics
 from rest.models import ClusterGroup, ClusterGroupRelation, Meeting, Metric, Node, Parameter, Record, RecordSet, Secret, SecretMeetingList, SecretMetricsList, Stats, RECORD_PROFILE_DESCRIPTION_LENGTH, MEETING_NAME_LENGTH
-from typing import Any, Dict, List, Literal
+from typing import Any, Dict, List, Literal, Union
 from urllib.parse import urlencode
 from xmltodict import parse
 
@@ -50,6 +50,7 @@ RETURN_STRING_VERSION = '<response>\r\n<returncode>SUCCESS</returncode>\r\n<vers
 RETURN_STRING_CREATE_LIMIT_REACHED = '<response>\r\n<returncode>FAILED</returncode>\r\n<message>Meeting/Attendee limit reached.</message>\r\n</response>'
 RETURN_STRING_CREATE_NO_NODE_AVAILABE = '<response>\r\n<returncode>FAILED</returncode>\r\n<message>No Node available.</message>\r\n</response>'
 RETURN_STRING_IS_MEETING_RUNNING_FALSE = '<response>\r\n<returncode>SUCCESS</returncode>\r\n<running>false</running>\r\n</response>'
+RETURN_STRING_GET_MEETING_INFO_FALSE = '<response>\r\n<returncode>FAILED</returncode>\r\n<messageKey>notFound</messageKey>\r\n<message>A meeting with that ID does not exist</message>\r\n</response>'
 RETURN_STRING_GET_RECORDING_TEXT_TRACKS_NOTHING_FOUND_JSON = '{"response":{"returncode":"FAILED","messageKey":"noRecordings","message":"No recording found"}}'
 RETURN_STRING_GET_RECORDING_NO_RECORDINGS = '<response>\r\n<returncode>SUCCESS</returncode>\r\n<recordings></recordings>\r\n<messageKey>noRecordings</messageKey>\r\n<message>There are no recordings for the meeting(s).</message>\r\n</response>'
 RETURN_STRING_MISSING_MEETING_ID = '<response>\r\n<returncode>FAILED</returncode>\r\n<messageKey>missingParamMeetingID</messageKey>\r\n<message>You must specify a meeting ID for the meeting.</message>\r\n</response>'
@@ -71,8 +72,8 @@ class ClientB3lbRequest:
     endpoint: str
     checksum: str
     stats_token: str
-    node: Node
-    secret: Secret
+    node: Union[Node, None]
+    secret: Union[Secret, None]
     state: str
     ENDPOINTS_PASS_THROUGH: List[str]
     ENDPOINTS: Dict[str, Any]
@@ -86,7 +87,7 @@ class ClientB3lbRequest:
         if not self.meeting_id:
             return HttpResponse(RETURN_STRING_MISSING_MEETING_ID, content_type=CONTENT_TYPE)
 
-        if not await sync_to_async(self.is_meeting)():
+        if not await self.is_meeting():
             if not await sync_to_async(self.is_node_free)():
                 return HttpResponse(RETURN_STRING_CREATE_NO_NODE_AVAILABE, content_type=CONTENT_TYPE)
             elif not await sync_to_async(self.is_in_limit)():
@@ -106,7 +107,7 @@ class ClientB3lbRequest:
         """
         if not self.meeting_id:
             return HttpResponse(RETURN_STRING_MISSING_MEETING_ID, content_type=CONTENT_TYPE)
-        if not await sync_to_async(self.is_meeting)():
+        if not await self.is_meeting():
             return HttpResponseBadRequest()
         await sync_to_async(self.check_parameters)()
         await sync_to_async(incr_metric)(Metric.JOINED, self.secret, self.node)
@@ -232,7 +233,7 @@ class ClientB3lbRequest:
         Checks meeting from B3LB database.
         Send client request to node and return response to client if exists otherwise return hardcoded answer.
         """
-        if not await sync_to_async(self.is_meeting)():
+        if not await self.is_meeting():
             return HttpResponse(RETURN_STRING_IS_MEETING_RUNNING_FALSE, content_type=CONTENT_TYPE)
         return await self.pass_through()
 
@@ -241,7 +242,12 @@ class ClientB3lbRequest:
         Multiple BBB endpoints.
         Send client request to correct node and return node response to client.
         """
-        self.set_node_by_meeting_id()
+        if not self.meeting_id:
+            return HttpResponse(RETURN_STRING_MISSING_MEETING_ID, content_type=CONTENT_TYPE)
+        if not await self.is_meeting():
+            return HttpResponse(RETURN_STRING_GET_MEETING_INFO_FALSE, content_type=CONTENT_TYPE)
+        if not self.node:
+            await self.set_node_by_meeting_id()
         async with ClientSession() as session:
             if self.request.method == "POST":
                 async with session.post(await sync_to_async(self.get_node_endpoint_url_encoded)(), data=self.body) as res:
@@ -420,9 +426,9 @@ class ClientB3lbRequest:
                 return False
         return True
 
-    def is_meeting(self) -> bool:
+    async def is_meeting(self) -> bool:
         if self.meeting_id:
-            self.set_node_by_meeting_id()
+            await self.set_node_by_meeting_id()
             if self.node:
                 return True
             return False
@@ -504,11 +510,11 @@ class ClientB3lbRequest:
         return dumps(statistic)
 
     ## Setter Routines ##
-    def set_node_by_meeting_id(self):
+    async def set_node_by_meeting_id(self):
         self.node = None
         if self.meeting_id:
             try:
-                meeting = Meeting.objects.get(id=self.meeting_id, secret=self.secret)
+                meeting = await sync_to_async(Meeting.objects.select_related("node").get)(id=self.meeting_id, secret=self.secret)
                 if not meeting.node.has_errors:
                     self.node = meeting.node
             except ObjectDoesNotExist:
@@ -531,7 +537,7 @@ class ClientB3lbRequest:
         if lowest_node_list:
             self.node = lowest_node_list[randint(0, len(lowest_node_list) - 1)]
 
-    def set_secret_by_slug_and_slug_id(self, slug: str, sub_id: int):
+    async def set_secret_by_slug_and_slug_id(self, slug: str, sub_id: int):
         if not slug:
             search = SLUG_REGEX.search(self.get_forwarded_host())
             if search:
@@ -539,7 +545,7 @@ class ClientB3lbRequest:
                 sub_id = int(search.group(3) or 0)
         if slug:
             try:
-                self.secret = Secret.objects.select_related("tenant", "tenant__asset").get(tenant__slug=slug.upper(), sub_id=sub_id)
+                self.secret = await sync_to_async(Secret.objects.select_related("tenant", "tenant__asset").get)(tenant__slug=slug.upper(), sub_id=sub_id)
             except ObjectDoesNotExist:
                 pass
 
@@ -582,7 +588,7 @@ class NodeB3lbRequest:
     """
     BACKENDS: Dict[str, Dict[Literal["methods", "function"], Any]]
     request: HttpRequest
-    meeting: Meeting
+    meeting: Union[Meeting, None]
     backend: str
     endpoint: str
     meeting_id: str
@@ -633,7 +639,7 @@ class NodeB3lbRequest:
     def full_endpoint(self) -> str:
         return f"{self.backend}/{self.endpoint}"
 
-    def get_record_set_by_nonce(self) -> RecordSet:
+    def get_record_set_by_nonce(self) -> Union[RecordSet, None]:
         try:
             record_set = RecordSet.objects.get(nonce=self.nonce)
         except ObjectDoesNotExist: